realtime_framework/realtime_data/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alpha-vantage==2.3.1
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.7
colorama==0.4.4
colorlog==6.8.2
comm==0.2.2
cryptography==35.0.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
docutils==0.18
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
feature-engine==1.8.0
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.3
importlib-metadata==4.8.1
ipykernel==6.29.4
ipython==8.25.0
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
keyring==23.2.1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.0.5
multitasking==0.0.11
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==1.26.4
overrides==7.7.0
packaging==22
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
peewee==3.17.5
pexpect==4.9.0
pkginfo==1.7.1
platformdirs==4.2.2
plumbum==1.7.0
polars==0.20.31
prometheus_client==0.20.0
prompt_toolkit==3.0.45
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pycparser==2.21
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
readme-renderer==30.0
redis==5.0.5
referencing==0.35.1
requests
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rpds-py==0.18.1
rpyc==5.0.1
rq==1.16.2
scikit-learn==1.5.0
scipy==1.13.1
SecretStorage==3.3.1
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.2
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4
tqdm==4.62.3
traitlets==5.14.3
twine==3.5.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.1
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.7
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
xgboost
yarl==1.9.4
yfinance==0.2.40
zipp==3.6.0
psycopg2-binary
python-dotenv
wandb
matplotlib
lightgbm
python-binance==1.0.19
